requirements.txt
fastapi==0.110.0
openai==0.27.2
PyJWT==2.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
bcrypt
python_jose==3.3.0
supabase==2.4.0
uvicorn==0.28.0
